Write attack method for computer Player
I wrote an attack method for the computer player
which makes random attacks on the opponents game
board. After a successful strike the method will
focus its attacks on the area of the successful first
strike.
I have yet to find a way to test this kind of method
where the ships of the opponent player are being
placed randomly on the game board, as well as the
other computer player first making random attacks
and then honing in on the target area.