Write a test to test the ComputerPlayer's attack method
I came up with a test to verify that the ComputerPlayer's
attack method focuses in on the area upon a first
strike. Therefore I placed a ship of known length
and position on a RealPlayer's game board and counted
the number of attacks needed by the ComputerPlayer
to sink the ship after a first strike (including
the first strike). The amount of attacks needed should
be less or equal to the attacked ship's length + 3 attacks.
The constant of '3' needs to be added since the computer
will make attacks in a breadth first fashion around the
first strike. This results in a maximum of 4 possibilities
(left, right, up, down). In its worst case only the last of
these 4 possibilities is successful and will need
another 'ship-length - 1' strikes to sink the ship.